+++ eddy/src/proofs.rs
@@ -17,7 +17,7 @@ impl TransparentEncryptionProof {
     }
 
     pub fn verify(&self, ctxt: &Ciphertext, encryption_key: &EncryptionKey) -> anyhow::Result<()> {
-        let limbs = Value::from(self.value as u64).to_limbs()?;
+        let limbs = Value::from(self.value).to_limbs()?;
         let ctxts = [ctxt.c0, ctxt.c1, ctxt.c2, ctxt.c3];
 
         for i in 0..4 {
